requirements.txt
# Ensure latest setuptools to avoid any package resources issues.
setuptools

nox
requests
pyyaml
jinja2

# some java processing requires xml handling
lxml

# Install sample-tester for generated samples
sample-tester